tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/005.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: document.write inline in markup </title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>log('inline script #1');
		document.write( '<script>log(\'doc write script\')<\/script>' );
		log('end script #1');
	</script>

	<script type="text/javascript">
	var t = async_test()

	function test() {
		assert_array_equals(eventOrder, ['inline script #1', 'doc write script', 'end script #1' ]);
		t.done();
        }
	onload = t.step_func(test)
	</script>

</body></html>
